pleasant pdf-generation form in progress
making a form with wtforms to slect background image for
form and generate pdf

added more responsive styling to word cards

made an "add" button for main word in word info page

gave the macro for add buttons a "infobutton=true"
argument so that the choice of not having the info button
is possible. Necessary in the wordinfo page for main word.